pyproject.toml
[project]
name = "bali-va"
version = "0.1.0"
description = ""
authors = [{ name = "Josh", email = "josh.yu_8@live.com" }]
requires-python = "~=3.10"
readme = "README.md"
dependencies = [
    "fastapi>=0.106,<1",
    "uvicorn>=0.29,<0.30",
    "sqlalchemy~=2.0",
    "pyhumps~=3.8",
    "fastapi-pagination>=0.12,<0.13",
]

[dependency-groups]
dev = [
    "pytest>=8.1.1,<9",
    "httpx>=0.27.0,<0.28",
]

[tool.hatch.build.targets.sdist]
include = ["bali"]

[tool.hatch.build.targets.wheel]
include = ["bali"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
